fix: allow specifying backend to collector
This is a subtly breaking change due to implicit argument order changes,
however there are not many users of the API who _don't_ use kwargs here.

Really it's a bit redundant to give a backend _and_ a tmp_dir and
ideally we wouldn't support both, but I don't want to change the public
API more than we have to.